filtering fixes for backends.
previously, snorkel was relying on the field type to be inside the field
name when running filters. now, fields can have any name, so we pass the
field type through to the filter row.

additionally, each adapter needs to be able to correctly create queries,
so now filters are: [filter, op, val] instead of [type_str.filter, op,
val] and the field type will be inside the query spec. the driver has to
appropriately name the field itself by using the column metadata,
instead of relying on the filter_row values to prefix the column types

test plan:

for each adapter in mongo, mongo_raw, postgres_raw and sybil:
  run queries with string/int filters through table popover and regular query interface