Use preset user in case there is an error setting up the socket
Sometimes, there is trouble looking up the user info in our session DB.
It's unclear why, this temporarily patches it.

The __user is mostly used inside the query controller and specifically
for keeping track of which users made which queries. In the failure
case, the personal query history and favorites will not be tracked.

This fixes #19

TEST PLAN: Couldn't repro original circumstance, so deliberately
commented out the lines setting the __user and verified that we
run into an error. Add the lines setting up default user and see
error go away.