Fix sporadic bug with compare queries in sybil driver
occassionally, a comparison query to sybil would send a float value for
the time column filter instead of an int, causing sybil backend to not
recognize the time filters.

the symptom is a comparison query result with 0 count (infrequently -
maybe 1 in 10 or so), when we know there actually is data available.